feat/chores: Serialization Behaviour, CI Improvements (#261)
## Description

This branch tightens serialization behavior and expands the validation
around it.

- Refines the runtime serializer and adds a dedicated failure exception
for clearer error handling.
- Adds broader serialization test coverage, including API contract,
exception contract, fuzz, and try-pattern tests.
- Updates runtime singleton and related helper code to support the
serialization changes.
- Expands pre-push, preflight, and CI automation for Unity, MCP,
documentation, and validation workflows.
- Refreshes supporting docs and repo configuration so the new checks and
workflows are documented and enforced consistently.

## Related Issue

Fixes #

## Type of Change

- [x] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [x] Documentation update
- [x] Refactor (code change that neither fixes a bug nor adds a feature)

## Checklist

- [ ] I have added tests that prove my fix is effective or my feature
works
- [ ] I have updated the documentation accordingly
- [ ] I have updated the [CHANGELOG](CHANGELOG.md)
- [ ] My changes do not introduce breaking changes, or breaking changes
are documented

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Hooks no longer run many checks locally (relying on `agent:preflight`
and CI), and Unity/self-hosted workflows are complex and
security-sensitive around licenses and log handling—misconfiguration
could block commits/pushes or leak diagnostics.
> 
> **Overview**
> **Git hooks** are restructured: shell entrypoints delegate to new
`.ps1` implementations and **require PowerShell 7**. **Pre-commit**
drops the old bash monolith (Prettier, cspell, markdownlint, CSharpier,
test lint, etc.) in favor of a **fast** path—version sync, final
newlines, LLM checks, staged `#region` guard, Unity `.meta`
handling—with guards against auto-staging files that already had
unstaged edits. Stray hook log artifacts are cleaned when gitignored.
**Pre-push** is similarly narrowed (notably **C# `#region` checks** on
pushed changes, with optional `agent:preflight:fix` path hints).
**Post-rewrite** clears the license-year cache via `git rev-parse
--git-path`.
> 
> **CI / GitHub Actions** gain substantial Unity support: composites for
**asmdef discovery**, **license preflight/return**, **result
verification** (NUnit XML, catastrophic log patterns, hang culprits),
**log tail dumps**, and **self-hosted runner diagnostics**; canonical
**`unity-versions.json`**, **`unity-test-project-modules.json`**, and
**`integration-packages.json`**. Smaller workflow tweaks: asmdef lint on
**`.cs`**, PR-scoped **license year** audits, **internal vs scheduled
external** doc links (lychee), **Windows + Ubuntu** LLM lint matrix, and
pwsh-invocation lint covering **actions** plus workflow
expression-length / Unity matrix contract tests. Misc: **actionlint**
self-hosted labels, docs URL placeholder update.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
cafaec16152781e6bca6e94cac3062929426e07c. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

---------

+++ Editor/Utils/WGroup/WGroupAnimationState.cs
@@ -29,7 +29,7 @@ namespace WallstopStudios.UnityHelpers.Editor.Utils.WGroup
         internal static AnimBool GetOrCreateAnim(
             WGroupDefinition definition,
             bool expanded,
-            int targetInstanceId = 0
+            long targetInstanceId = 0
         )
         {
             int key = ComputeKey(definition, targetInstanceId);
@@ -60,7 +60,7 @@ namespace WallstopStudios.UnityHelpers.Editor.Utils.WGroup
         internal static float GetFadeProgress(
             WGroupDefinition definition,
             bool expanded,
-            int targetInstanceId = 0
+            long targetInstanceId = 0
         )
         {
             if (!UnityHelpersSettings.ShouldTweenWGroupFoldouts())
@@ -89,7 +89,7 @@ namespace WallstopStudios.UnityHelpers.Editor.Utils.WGroup
             FoldoutAnimations.Clear();
         }
 
-        private static int ComputeKey(WGroupDefinition definition, int targetInstanceId)
+        private static int ComputeKey(WGroupDefinition definition, long targetInstanceId)
         {
             return Objects.HashCode(
                 definition.Name,
